% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_transitions.R
\name{boot_transitions}
\alias{boot_transitions}
\title{Bootstrap observed census transitions}
\source{
This is a modified version of \code{\link[popbio]{boot.transitions}}.
}
\usage{
boot_transitions(transitions, iterations, by.stage.counts = FALSE, ...)
}
\arguments{
\item{transitions}{a stage-fate data frame with stage or age class in the current census, fate in the subsequent census, and one or more fertility columns}

\item{iterations}{Number of bootstrap iterations}

\item{by.stage.counts}{Resample transitions with equal probability (default) or by subsets of initial stage counts}

\item{...}{additional options passed to \code{\link[raretrans]{projection_matrix}}}
}
\value{
A list with 3 items

lambda	A vector containing bootstrap values for lambda

matrix	A matrix containing bootstrap transtion matrices with one projection matrix per row.

vector	A matrix containing bootstrap stage vectors with one stage vector per row.
}
\description{
Calculate bootstrap distributions of population growth rates (lambda), stage vectors, and projection matrix elements by randomly sampling with replacement from a stage-fate data frame of observed transitions
}
