% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\name{document_data}
\alias{document_data}
\title{Add roxygen2 documentation template for a dataset}
\usage{
document_data(object)
}
\arguments{
\item{object}{data set to document}
}
\description{
Add roxygen2 documentation template for a dataset
}
