% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_funcs.R
\name{fill_fecundity}
\alias{fill_fecundity}
\title{Combine prior and data for fertility matrix}
\usage{
fill_fecundity(TF, P, N, returnType = "A")
}
\arguments{
\item{TF}{A list of two matrices, T and F, as ouput by \code{\link[popbio]{projection.matrix}}.}

\item{P}{A matrix of the priors for each column.}

\item{N}{A vector of observed transitions.}

\item{returnType}{A character vector describing the desired return value.}
}
\value{
The return value depends on parameter returnType.
\itemize{
  \item A - the summed matrix "filled in" using a dirichlet prior
  \item F - just the filled in fertility matrix
}
}
\description{
\code{fill_fertility} returns the expected value of the fertility
matrix combining observed recruits for one time step and a prior
}
\details{
This is just an intermediate version to test package structure. Does
not yet use the prior argument P.
}
