% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_matrix.R
\name{get_state_vector}
\alias{get_state_vector}
\title{Helper functions for generating different priors}
\usage{
get_state_vector(transitions, stage = NULL, sort = NULL)
}
\arguments{
\item{transitions}{a dataframe of observations of individuals in different stages}

\item{stage}{the name of the variable with the stage information}

\item{sort}{a vector of stage names in the desired order. Default is the order of levels in stage.}
}
\value{
a vector of the counts of observations in each level of stage.
}
\description{
Extract the number of individuals in each stage from a dataframe
of transitions.
}
\examples{
data("L_elto")

# Extract one population at one census period
onepop <- L_elto[L_elto$POPNUM == 250 & L_elto$year == 5, ]
onepop$stage <- factor(onepop$stage, levels = c("p", "j", "a"))

# Count individuals per stage in the order p, j, a
get_state_vector(onepop, stage = "stage", sort = c("p", "j", "a"))
}
