% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_matrix.R
\name{projection_matrix}
\alias{projection_matrix}
\title{Construct projection matrix models using transition frequency tables}
\source{
This is a modified version of \code{\link[popbio]{projection.matrix}}.
}
\usage{
projection_matrix(
  transitions,
  stage = NULL,
  fate = NULL,
  fertility = NULL,
  sort = NULL,
  add = NULL,
  TF = FALSE
)
}
\arguments{
\item{transitions}{a stage-fate data frame with stage or age class in the current census, fate in the subsequent census, and one or more fertility columns}

\item{stage}{a column name or position of the stage column in the stage-fate data frame. Defaults to "stage".}

\item{fate}{name of the fate column in the stage-fate data frame. Defaults to "fate". Individuals that die can be indicated with any character string not in the levels of stage, or missing.}

\item{fertility}{one or more names of fertility columns in the stage-fate data frame. By default, any column names matching stage class names are assumed to contain individual fertilities}

\item{sort}{a vector listing stage classes that correspond to the rows and columns of the desired projection matrix. Currently, names in this vector must match a level in the stage column. Also, this option should only be used if stages are not ordered, since the default is to sort by levels in the stage column.}

\item{add}{a vector listing row, column and value, used to add estimated transtions to the transition matrix (e.g., a transition from seed bank to seedling). May be repeated.}

\item{TF}{output separate transition (T) and fertility (F) matrices. Default is FALSE and outputs a single projection matrix A

The state transition rates are estimated using transition frequency tables (see section 6.1.1, Caswell 2001), so this technique will most likely apply to demographic studies of plants or other sessile organisms where individuals are tagged and then consistently relocated in annual censuses. The fertility rates are calculated by averaging individuals fertilities by stage class; therefore, some care should be taken to correctly estimate individual fertilities based on the timing of the census.
Individual fertilities should be the total number of offspring at the end of the census interval. Therefore, fertilites should include offspring survival in a prebreeding censuses (and more than one offspring class may be present). In a postbreeding census, new offspring were born just before the census, so the fertility rate is just the number of offspring in this case.}
}
\value{
The default output is a single projection matrix A. If the TF flag is true, then a list with 2 items where A=T+F
}
\description{
Construct an age or stage-structure projection model from a transition table listing stage in time t, fate in time t+1, and one or more individual fertility columns.
This version is modified to allow for missing transitions in the data.
}
