% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raretrans.R
\docType{package}
\name{raretrans}
\alias{raretrans}
\alias{raretrans-package}
\title{raretrans: A package for creating and updating Bayesian priors for matrix population models}
\description{
This package provides functions that combine data on transitions with prior beliefs about
transition probabilities and fertilities.
}
