% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the interactive Shiny application for Bayesian transition priors}
\usage{
run_app()
}
\value{
No return value, called for side effects (launches a Shiny app).
}
\description{
Run the interactive Shiny application for Bayesian transition priors
}
\examples{
if(interactive()){
raretrans::run_app()
}
}
