% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funcs.R
\name{sim_observations}
\alias{sim_observations}
\title{Simulate observations from a given matrix}
\usage{
sim_observations(TF, N = 100, samples = 1)
}
\arguments{
\item{TF}{A list of two matrices, T and F, as ouput by \code{\link[popbio]{projection.matrix}}.}

\item{N}{integer. Number of plants to assume in the population.}

\item{samples}{integer. Number of observations to generate from matrix.}
}
\value{

}
\description{
Simulate observations from a given matrix
}
