% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_funcs.R
\name{fill_fecundity}
\alias{fill_fecundity}
\title{Combine prior and data for fertility matrix}
\usage{
fill_fecundity(TF, N, alpha = 1e-05, beta = 1e-05, priorweight = -1,
  returnType = "A")
}
\arguments{
\item{TF}{A list of two matrices, T and F, as ouput by \code{\link[popbio]{projection.matrix}}.}

\item{N}{A vector of observed stage distribution.}

\item{alpha}{A vector of the prior parameter for each stage. Stages that can't reproduce are NA_real_}

\item{beta}{A vector of the prior parameter for each stage. Stages that can't reproduce are NA_real_}

\item{priorweight}{total weight for each column of prior as a percentage of sample size or 1 if negative}

\item{returnType}{A character vector describing the desired return value.}
}
\value{
The return value depends on parameter returnType.
\itemize{
  \item A - the summed matrix "filled in" using a Gamma prior
  \item F - just the filled in fertility matrix
}
}
\description{
\code{fill_fertility} returns the expected value of the fertility
matrix combining observed recruits for one time step and a Gamma prior for each column.
}
\details{
Assumes that the reproductive individuals are in stage 3 ... needs generalizing.
}
