% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_funcs.R
\name{fill_fertility}
\alias{fill_fertility}
\title{Combine prior and data for fertility matrix}
\usage{
fill_fertility(TF, N, alpha = 1e-05, beta = 1e-05, priorweight = -1,
  returnType = "F")
}
\arguments{
\item{TF}{A list of two matrices, T and F, as ouput by \code{\link[popbio]{projection.matrix}}.}

\item{N}{A vector of observed stage distribution.}

\item{alpha}{A matrix of the prior parameter for each stage.}

\item{beta}{A matrix of the prior parameter for each stage.}

\item{priorweight}{total weight for each column of prior as a percentage of sample size or 1 if negative}

\item{returnType}{A character vector describing the desired return value. Defaults to "F" the fertility matrix}
}
\value{
The return value depends on parameter returnType.
\itemize{
  \item A - the summed matrix "filled in" using a Gamma prior
  \item F - just the filled in fertility matrix
  \item ab - the posterior parameters alpha and beta as a list.
}
}
\description{
\code{fill_fertility} returns the expected value of the fertility
matrix combining observed recruits for one time step and a Gamma prior for each column.
}
\details{
Assumes that only one stage reproduces ... needs generalizing.
}
