% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_funcs.R
\name{fill_transitions}
\alias{fill_transitions}
\title{Combine prior and data for transition matrix}
\usage{
fill_transitions(TF, N, P = NULL, priorweight = -1, returnType = "A")
}
\arguments{
\item{TF}{A list of two matrices, T and F, as ouput by \code{\link[popbio]{projection.matrix}}.}

\item{N}{A vector of observed transitions.}

\item{P}{A matrix of the priors for each column. Defaults to uniform.}

\item{priorweight}{total weight for each column of prior as a percentage of sample size or 1 if negative}

\item{returnType}{A character vector describing the desired return value.}
}
\value{
The return value depends on parameter returnType.
\itemize{
  \item A - the summed matrix "filled in" using a dirichlet prior
  \item T - just the filled in transition matrix
  \item TN - the augmented matrix of fates -- use in calculating the CI or for simulation
}
}
\description{
\code{fill_transition} returns the expected value of the transition
matrix combining observed transitions for one time step and a prior
}
