% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_funcs.R
\name{get_state_vector}
\alias{get_state_vector}
\title{Extract the number of individuals in each stage from a dataframe
of transitions}
\usage{
get_state_vector(transitions, stage = NULL, sort = NULL)
}
\arguments{
\item{transitions}{a dataframe of observations of individuals in different stages}

\item{stage}{the name of the variable with the stage information}

\item{sort}{a vector of stage names in the desired order. Default is the order of levels in stage.}
}
\value{
a vector of the counts of observations in each level of stage.
}
\description{
Extract the number of individuals in each stage from a dataframe
of transitions
}
