% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_elto.R
\docType{data}
\name{L_elto}
\alias{L_elto}
\title{transition and recruitment data for \emph{L. elto}}
\format{A data frame with 5233 rows and 13 variables}
\source{
Raymond L. Tremblay, University of Puerto Rico, Unpublished data.
}
\usage{
data(L_elto)
}
\description{
6 month periodic census of several populations of the epiphytic Orchid \emph{Lepenthes elto}.
}
\details{
\itemize{
   \item POPNUM...
   \item year...
   \item seedlings...
   \item adults...
   \item fertility...
   \item IND_NUM...
   \item stage...
   \item next_stage...
   \item first_year...
   \item last_year...
   \item recruited...
   \item died...
   \item lifespan...
 }
}
\references{
{
Tremblay, Raymond L., and Michael J. Hutchings. "Population dynamics in orchid conservation: a review of analytical methods based on the rare species Lepanthes eltoroensis." Orchid Conservation, Borneo: Natural History Museum Publications (2003): 183-204.
}
}
\keyword{...}
