% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raretrans.R
\docType{package}
\name{raretrans-package}
\alias{raretrans}
\alias{raretrans-package}
\title{raretrans: Bayesian Priors for Matrix Population Models}
\description{
Provides functions to correct biased transition and fertility estimates in
population projection matrices caused by small sample sizes. Small or
short-term studies frequently produce structural zeros (biologically possible
transitions never observed) and structural ones (transitions estimated at
100% survival, stasis, or mortality that are biologically implausible).
Both distort matrix structure and bias estimates of population growth.
Functions combine observed transition data with Bayesian prior beliefs to
regularise estimates from rare or unobserved events.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://atiretoo.github.io/raretrans/}
  \item \url{https://github.com/atiretoo/raretrans}
  \item Report bugs at \url{https://github.com/atiretoo/raretrans/issues}
}

}
\author{
\strong{Maintainer}: Raymond Tremblay \email{raymond.tremblay@upr.edu} (\href{https://orcid.org/0000-0002-8588-4372}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Raymond Tremblay \email{raymond.tremblay@upr.edu} (\href{https://orcid.org/0000-0002-8588-4372}{ORCID}) [copyright holder]
  \item Andrew Tyre \email{atyre2@gmail.com} (\href{https://orcid.org/0000-0001-9736-641X}{ORCID}) [copyright holder]
  \item Maria-Eglee Perez \email{mariaeglee@gmail.com} (\href{https://orcid.org/0000-0001-8641-8405}{ORCID})
}

Other contributors:
\itemize{
  \item Sasha Tenhumberg [contributor]
}

}
