% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funcs.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Dirichlet distributed random numbers}
\source{
copied from package \code{MCMCpack} to avoid a dependency. That code
was taken from Greg's Miscellaneous Functions (gregmisc). His code was based
on code posted by Ben Bolker to R-News on 15 Dec 2000.
}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{The number of random vectors to generate}

\item{alpha}{A vector of parameters}
}
\value{
The function returns a matrix with n rows and \code{length(alpha)} columns
}
\description{
Generate n random vectors distributed according to a Dirichlet distribution.
Each row of the returned matrix is a random vector that sums to 1.
}
